Move tintaConfigDir to its own translation unit
The i18n and inline-style test targets compile i18n.cpp and themes.cpp,
which call tintaConfigDir since portable mode (#151) but do not link
settings.cpp - the v3.4.0 tag build failed at link. The resolver now
lives in config_dir.cpp, linked by the app and both test targets.
